# Conflicts:
#	server/controllers/locks_controller_test.go
#	server/core/locking/locking.go
#	server/core/locking/mocks/mock_backend.go

+++ server/events/working_dir.go
@@ -14,19 +14,24 @@
 package events
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
+	"sync"
 
 	"github.com/pkg/errors"
 	"github.com/runatlantis/atlantis/server/events/models"
 	"github.com/runatlantis/atlantis/server/logging"
+	"golang.org/x/sync/semaphore"
 )
 
 const workingDirPrefix = "repos"
+
+var cloneLocks sync.Map
 
 //go:generate pegomock generate -m --use-experimental-model-gen --package mocks -o mocks/mock_working_dir.go WorkingDir
 //go:generate pegomock generate -m --use-experimental-model-gen --package events WorkingDir
@@ -66,13 +71,15 @@ type FileWorkspace struct {
 	// from the "origin" remote. If this is false, we fetch "+refs/heads/$HEAD_BRANCH"
 	// from the "head" remote.
 	GithubAppEnabled bool
+	// use the global setting without overriding
+	GpgNoSigningEnabled bool
 }
 
 // Clone git clones headRepo, checks out the branch and then returns the absolute
 // path to the root of the cloned repo. It also returns
 // a boolean indicating if we should warn users that the branch we're
 // merging into has been updated since we cloned it.
-//If the repo already exists and is at
+// If the repo already exists and is at
 // the right commit it does nothing. This is to support running commands in
 // multiple dirs of the same repo without deleting existing plans.
 func (w *FileWorkspace) Clone(
@@ -197,6 +204,18 @@ func (w *FileWorkspace) forceClone(log logging.SimpleLogging,
 	headRepo models.Repo,
 	p models.PullRequest) error {
 
+	value, _ := cloneLocks.LoadOrStore(cloneDir, semaphore.NewWeighted(1))
+	sem := value.(*semaphore.Weighted)
+
+	defer sem.Release(1)
+	if acquired := sem.TryAcquire(1); !acquired {
+		err := sem.Acquire(context.TODO(), 1)
+		if err != nil {
+			return errors.Wrap(err, "waiting for repository to be cloned")
+		}
+		return nil
+	}
+
 	err := os.RemoveAll(cloneDir)
 	if err != nil {
 		return errors.Wrapf(err, "deleting dir %q before cloning", cloneDir)
@@ -240,16 +259,21 @@ func (w *FileWorkspace) forceClone(log logging.SimpleLogging,
 			{
 				"git", "fetch", fetchRemote, fetchRef,
 			},
-			// We use --no-ff because we always want there to be a merge commit.
-			// This way, our branch will look the same regardless if the merge
-			// could be fast forwarded. This is useful later when we run
-			// git rev-parse HEAD^2 to get the head commit because it will
-			// always succeed whereas without --no-ff, if the merge was fast
-			// forwarded then git rev-parse HEAD^2 would fail.
-			{
-				"git", "merge", "-q", "--no-ff", "-m", "atlantis-merge", "FETCH_HEAD",
-			},
 		}
+		if w.GpgNoSigningEnabled {
+			cmds = append(cmds, []string{
+				"git", "config", "--local", "commit.gpgsign", "false",
+			})
+		}
+		// We use --no-ff because we always want there to be a merge commit.
+		// This way, our branch will look the same regardless if the merge
+		// could be fast forwarded. This is useful later when we run
+		// git rev-parse HEAD^2 to get the head commit because it will
+		// always succeed whereas without --no-ff, if the merge was fast
+		// forwarded then git rev-parse HEAD^2 would fail.
+		cmds = append(cmds, []string{
+			"git", "merge", "-q", "--no-ff", "-m", "atlantis-merge", "FETCH_HEAD",
+		})
 	} else {
 		cmds = [][]string{
 			{
